Change NavigationView property 'animated' to 'pageAnimation'
The boolean property 'animated' did not provide enough future
flexibility to expand the animation behavior. We therefore change the
property 'animated' to 'pageAnimation' which currently accepts the two
values 'default' and 'none'. It thereby provides the same functionality
as the 'animated' did but allows for different animations in the future.

Change-Id: I6c4b115373f41329c2a2590862eb742ff379d53e