[TypeScript] Improve type definition for Widget.apply
Introduce a "ts_type" field for method parameters in the JSON doc files
for TypeScript specific type information.

The new type for the parameter of Widget.apply allows for autocompletion
and type information for properties that are defined on Widget while
still allowing to change properties that are not defined on Widget.

Change-Id: I454666e32e02d0cd7c7badb12a419ef9ac4e4f75