Revert global flushed event
Revert the fix made in 5207a40 (introducing the 'flushed' event).

The problem this change addressed (#1291) exists only on Picker,
which sets native *properties* in the flush listener instead of
calling native methods. Properties need to be flushed explicitly
at this point, as there will be no subsequent flush.

Work around this problem in Picker for now, think about a better
approach to the general problem of flushing the native bridge
later on.

See also #1291

Change-Id: I80662e636f0faa70b3c00b020ece5b31366fdd75