Add windows-only property "win_tabBarTheme" to TabFolder
This is consistent with `NavigationView.win_toolbarTheme` and allows a
workaround for `textColor` not being supported on windows. It is
currently unclear if there is a good way to support textColor on
windows. Also, this property is useful either way since it allows using
the different native windows theme colors for the bar, which is more
common than applying a custom color.

Change-Id: I29bcffe213aa21b90d0ef74fce3ea738ce4f936c