Modernize Bookstore example
Migrate example to ES 6 and split the file bookstore.js into several
classes to improve comprehension.

Extract strings to constants to make the code easier to read.

Create the UI in a more structured manner to be able to make UI hierarchy
more apparent and to separate the concerns of applying layout and
styles.

Change-Id: I53ae70fc4d2fc3eb6d8b87974456e498462f56b1