Introduce basic filesystem API
Existing Cordova plug-ins for file access are based on the W3C File
APIs which are dated and complicated. They also rely on Blob for
binary data which is not available in Tabris.

Provide a simple API to read, write, and delete files. Implement the
methods asynchronously and return promises to avoid blocking the UI
thread by file operations.

In a second step, we should add methods for dealing with directories,
handling text encodings etc.

Instead of using events, try a new approach and pass callbacks to the
clients which will then be called. This should be simpler and faster
than using the legacy JSON protocol operations. If this works out, we
can extend this approach to other types.

See also #1364

Change-Id: Ifb69105fbb8c4f58fdbba513d8a7ea173e9f651f