Incorporate polyfills for Promise and Object.assign
The polyfill for Object.assign() will be removed in a future version
when all clients implement this method.

The Promise polyfill is still required, as Promises must be based
on the native setTimeout function to call back into the JS context.
However, the current polyfill is based on asap, a wrapper around
browser functions such as setImmediate which are not available in
Tabris.js.

Incorporate the relevant parts of the polyfills for Object.assign
and Promise and adjust the code to ES6.
This removes the need for a separate polyfill.js file and also for
webpack and uglify in the build chain.

Do not add the Promise method done() that was available before but is
not part of the ES2015 spec.

Change-Id: I54eacbb5492872fccca960a2d72527ab2fa425b8