Prevent loading tabris twice
When an app depends on modules that in turn depend on tabris, those
modules can end up with different versions of the tabris module.
Only one of these tabris modules would be eventually initialized by
the native client. This setup would lead to incomprehensible errors.

Ensure that the tabris module is not already installed by checking
the global tabris object. Distinguish this object from provisional
tabris object installed during the bootstrap process by checking
both the object and its version property.

Add a test that creates a copy of the tabris module and verifies
that it cannot be required.

Fix #1380

Change-Id: I32a6a289203009b2afb5461cf3bf1c8288e201f7