Call a native method when adding widgets to parent
When widgets are added to a parent, only the native property `parent`
is being set. Hence, the native clients do not receive the insertion
index when widgets are added using insertBefore() or insertAfter().

Call a native `insert()` method on the parent, with the added child and
the insertion index as parameters. Keep the parent property for now to
ease the migration. When clients have adjusted, it can be removed.

See also #1224

Change-Id: I186dc55a730cb62c2fd477756e7a2442a7917ee7