Translate null to '' for string properties
The property encoding function for strings currently translates all
values to strings. This includes translating `null` to the string
"null", which has been found confusing. When setting a text property
to `null`, the intention is rather to reset the property.

Translate `null` to the empty string instead of "null".

Keep the existing behavior (translating `null` to "null") for the
storage API.

Fix #996

Change-Id: I5e34e142addb2690ac1488c5f5a59e6a88848f0a