Translate undefined to null in native bridge
Some property encoders return `undefined` to remove a property from
the cache (font, color, see aa0d6fa). In these cases, the value must
be transferred to the client as `null`.

Since the native clients never expect `undefined` as a value,
generally translate `undefined` to `null` in the default setter.
This establishes a standard way of resetting properties by returning
`undefined` from the encoder without the need for a custom setter.

Fix #1351

Change-Id: Ia7667ad5a84a384114e4e21a13b9af6da66e69a7