Make version check more strict
As of Tabris 2.0, the native client's version must match the version
of the tabris module. This allows us to evolve the protocol during
the release stream. Patch versions should not be ignored as they
contain only bug fixes and don't affect the protocol.

The CLI will download the native client and so ensure compatible
versions.

With this change, major and minor segments must match, so it's now
sufficient to compare the segments as strings.

Change-Id: I6e7a71dd045816c6b8f1f5402f822e3c1fdd972e