Allow using constructors as selectors
All methods that previously supported strings and filter functions as
selectors now also support widget constructors. To distinguish between
filter and constructor functions the prototype property of the function
is checked. In is expected to have a prototype chain that ends in the
NativeObject prototype. Using the Widget prototype did not work since
this creates a circular dependency and crashes when GestureRecognizer
calls NativeObject.defineProperties. 

Extend and improve selector type documentation. 

Adjust typescript declaration generator to allow generics and
TypeScript-only method signatures. Unlike the markdown documentation the
TypeScript declarations have to distinguish between the constructors and
other selectors to determine the generic type.

Adjust input.tsx example to use this feature. 

Change-Id: I5f782a8601ce0e0846067ea9061478e6ed9938ad