Fix resolution of canvas examples
The width and height arguments supplied to getContext dictate the 
size of the render target. In Tabris.js 2 the render target is 
stretched to fit onto the canvas, not clipped. For this reason, 
unless super- or subresolution is intended, the dimension of the 
context should always be the dimensions of the canvas multiplied 
with the device scale factor.

Change the canvas animation example to only update angle and fps when
animating.

Change-Id: Iad60cc5af2da31dec8a4c97f7fcc19aa94bd068d