[doc] Revise modules article
Rename sections to "Importing Modules" and "Module Exports" as these
are the important parts of a module system.

Move the paragraphs on best practices to the beginning of the
"Module Exports" section.

Add a link to the TypeScript handbook article about modules.

Remove explanation on loading files from the local device in remote
mode. It turns out that files are never loaded from the device due
to a different path - and the actual behavior is actually preferable.

Use single quotes in code.

Change-Id: Iad6ada6fb0ac51b8ceaa6b76b7017bd43c2a6265