Fix generation of change events
Generate change events as `propertyChanged` instead of the obsolete
`change:property`.

Align description of generated change events with the tense in the
property name to "... when the property has changed".

Only add generated change events where the event is not part of the
declaration to prevent duplicate definitions in the doc. We might
sort out these cases later.

Change-Id: If87daf243f4d70331ee67ac28b811ad2465d8ea0