Allow documenting global API with json API files
This allows a key 'namespace' to the root of the json API definition
files that mark the contained API (e.g. methods) as globally available.
Global types are not yet supported. The `type` field is only used to
name the definition. It should not be needed in the future. 

Add a single line on top of the generated files that describe how to
access the API (import or globally available).

Rename the winndow.json to timer.json and mark it as global.

Change-Id: Iacce64c383391209f251163e01ef486c9dab8502