[doc] Replace info icon with emoji
Be consistent and replace the info icon with the :point_right: emoji
everywhere. The info icon lead to problems with indentation, emojis
give us more freedom to express different kind of hints in the future.

Remove the surrounding <i> element from quotes. It conflicts with
*emphasis* within the markdown. If really needed, the text style
should be set in the CSS which is now under our control.

No content changes.

Change-Id: I7f0480b6561cb309ffcbcb419b7f3ccb78f706f5