[doc] Rework doc generator to not require a "type" field anymore
The documents title is now determined by the field "title", "object" or
"type", in that order.

The former types array is now a typeLinks map containing all known type
links generated in advance. The api map can then use the final title of
the document as a key instead of the type. 

Some minor side effects of this change:
 - Case insensitivity regarding type to type link mapping is lost.
However, none of the JSON files seem to have relied on that "feature".
 - Snippet lookup now uses the title rather than the type. This results
in the 'fs.js' snippet to be picked up, which it wasn't before.

Change-Id: I9e9913bb37415a687788b82edc40c9c9b1c1c5a6