Bundle global typings in a single file
Type definitions that are not in the tabris namespace and not generated
by the type definition generator are stored in seperate files in the
`typings` directory.

Concatenate all global typings into a single `globals.d.ts` type
definition file during build instead of copying and referencing all
files seperately.

Change type definitions for JSX to work when concatenated with the other
type definition files.


Change-Id: I27a18a20d5bef0212c4c2a1323d3f305d3671c12