Support padding on Composite
Introduce padding as a new property on Composite.

This property is independent of layout data, it's not a part of the
layout, but of the native widgets.

Introduce a new property type box dimension - an object with optional
members for the edges: left, right, top, bottom.

Also allow to set the padding to a single number that would be applied
on all four edges.

See also #1017

Change-Id: I24e10df98ab8b7a661a13289d931e6bad7f8f476