Implement console in JavaScript
Implement the console object [1] in JavaScript.

This ensures that different types of values are printed consistently
across platforms. This implementation also supports printf-like
placeholders, as the console in Browser and Node.js does.

Implement a Formatter class that contains a stripped-down version of
the formatting code from Node.js [1].

Replace the existing console if and only if the native console object
provides a print function. Platforms that don't provide this function
are not affected.

[1] https://github.com/nodejs/node/

See also #1459

Change-Id: I38b23660f1d086b10b88571eae14dbe26e77ec21