Decouple build scripts from grunt
The common structure of Grunt files separates build tasks from their
configuration. As an effect, information is distributed between the
config objects and the tasks itself, which makes these files very hard
to understand. See also [1].

To open up the option of replacing Grunt with a simpler tool, replace
the dependencies to Grunt API in the external scripts with common node
API.

[1] https://www.keithcirkel.co.uk/why-we-should-stop-using-grunt/

Change-Id: I463c32bad22550eb44c0c2994fabe43c418e9d4b