Use uglify-es to minify JavaScript
We moved away from uglify in 63f19b78 because of its lack of ES6
support and mangling of constructor names, and switched to an early
version of the babel minifier (babili@0.0.11).

Newer versions of babili lead to unexplainable "xxx is not defined"
errors, so we have to look for an alternative. Babili also has the
disadvantage that it minifies during transpilation, and so does not
produce an intermediate unminified file.

The uglify-es module [1] is created from the harmony branch of uglify
and it fits our needs now. The compression is also better than with
babili (128k vs 132k).

[1] https://www.npmjs.com/package/uglify-es

Change-Id: I14479813b9e12352a194be3104446e18867d6a74