Introduce TextEncoder
Analogous to TextDecoder, introduce a native TextEncoder that
encodes string to array buffers. This is a precondition for
support for writing text files.

Export a class with a static decode method instead of just exporting
a method in order to be able to stub the method in tests. Apply the
same pattern to TextDecoder.

Change-Id: Ic4eea2248247219298ad19cfb710db38cd9554de