Add support for text files to the filesystem API
When the second argument to writeFile is a string, encode this string
before writing to file, using the encoding specified in the additional
optional parameter `encoding`, or UTF-8 if not set.

Also add an additional optional parameter `encoding` to readFile.
When this parameter is set, decode the contents and return the string.

This API is very close to the Node.js file API except that it returns
a promise instead of accepting a callback. Node.js also supports an
options object instead of the encoding string, this can still be added
when needed.

Change-Id: Ibcd8062e58159f28d89c6050ee93a1c12d0ff4e4