Use low level console.print() method in boot.js
While bootstrapping, the console object is not yet replaced by the
tabris module. Do not rely on the `console.error()` method in the
bootstrap code. Use the native console.print() method instead.

With this change, the native platforms don't need to implement
console.debug(), .log(), info(), .warn(), and .error() anymore.

Change-Id: Ia17a4f72c1aad6124fa741b712a3970b070013f3