[doc] Escape property types containing a pipe
The documentation contains a pipe character ("|") to denote multiple
types that can be set or returned for a property.  This is also the
character that Markdown uses to denote the boundaries between table
cells.  Markdown only uses the separator between cells, and not at the
start of a row, so the string "number|string" would be parsed as a table
row.  Instead, split the types by a pipe, render links as appropriate,
and then join the types together with an escaped pipe to prevent table
rendering.