Create a common base class for all Pop-ups
The "Popup" class is now the base class for ActionSheet and 
AlertDialog. It may also be used for classes like Menu or various 
dialogs/picker in the future. It only implements the open/close 
methods.

The rational for this change is that until recently
the rule was that all visual elements were based on Widget. 
AlertDialog was an exception, as it did not have enough commonality 
with widget. Now that ActionSheet has the same issue, the rule can be
consolidated as "all visual elements are widgets or pop-ups".
 
Change-Id: I9539fb21c347150fc9a6b3ae7be9338720566c8b