[TypeScript] Remove type parameter from events interfaces
Previously, the `WidgetEvents` interfaces had a type parameter in order
for the `event.target` property to be the type of the widget the event
was called on.

Remove the type parameter and use the type the event was defined on for
`event.target`.

This simplifies the events interfaces with minimal loss of functionality 
and removes the necessity of future event object interfaces to be generic.

Change-Id: Ifbf9a38a034949b3039555ca17f7b1b26dd9fc7b