[TypeScript] Generate EventObject interfaces
Generate interfaces for the event object of any event using the
'<Widget><EventName>Event' naming scheme where <EventName> is the
capitalized name of the Event and <Widget> is the name of the Widget the
event is defined on (i.e. `TextInputAcceptEvent`). Events, that have no
additional parameter use the `EventObject` interface. Where appropriate,
a single event object interface is used for multiple events. For
example, `panUp`, `panDown` etc. all use the `WidgetPanEvent`.

Remove index parameters on both EventObjects and EventsObjects.
This means, that adding arbitrary properties to event objects or adding
listener to arbitrary events in the `on` method using object notation,
is no longer possible.

Change-Id: Ia300861d77897609be46e50e79ca77f90442f369