Use internal console log functions
The global console object can not reliably be assumed to be the wrapped
console object with all the individual logging functions, it may also be
the native console object that only has the "print" function. 

Let the Console object export individual log functions to use internally
that are guaranteed to work at all times - even if the application code
messes with the console object.

In tests the functions simply redirect to the native console directly.

Change-Id: I9c9b01a9e22a8a0a5ab4561f5843b6dad41032e9