Make EventObject a public class
This change makes it simpler for application developers to trigger
custom events that have the same interface as built-in events.
EventObject can be constructed without parameters, and the properties
will then be set by the `trigger` method. Other objects are treated as
before.

Revise documentation accordingly. EventObject is now also documented
with a .json file. This makes the explicitly defined TypeScript
"EventObject" interface obsolete.

The internal mechanism that automatically attaches custom properties to
an eventObject has been moved form EventObject to NativeObject.

Change-Id: If7cf66b29e59618a47f6cda2cdc68539b97a70d9