Add DateDialog to pick a date
The DateDialog allows the user to select a date from a modal dialog.
It can be initialized with a starting date as well as a date range from
which the user can pick a date.

Listeners can be attached to be informed of the selected date or whether
date selection has been canceled.

Change-Id: Ia1acbb4183ebd177b2470509da5b0e223ed753e9