Allow string literals inside JSX elements
Part of an ongoing effort to leverage the full feature set of JSX. See
https://reactjs.org/docs/jsx-in-depth.html#string-literals-1

Previously this just caused an exception. Now the given string will be
set if the widgets defines a "text" property. If there is none the
string is ignored and a warning is printed. This is an existing feature
of NativeObject.

Change-Id: Iace3c70feb23cb160a4d2150927e56ee6862d386