[TypeScript] Do not override `set` in every NativeObject class
Instead of overriding set in every class that extends NativeObject, use
a special typescript-only property `tsProperties` to reference the
property map in a generic way. Also use this approach to define
constructors.

Additional benefits:
 - Type safety extends to `set(key, value)` and `get(key)`
 - Smaller tabrid.d.ts

Revise and extend lang.md to explain in detail how to use
this feature. Revise input-tsx example to demonstrate this feature in
almost all variations.

Change-Id: I08c57dacf88da115851b49323ef8e6c352cc82cc