Disable tsconfig flag "strictFunctionTypes"
This flag was introduced in TypeScript 2.6 and was implicitly enabled
via the "strict" flag that is set in the tsconfig of the typescript
tests. Unfortunately it conflicts with the way some listener parameter
objects are declared. 

For now, replace the "strict" flag with the individual flags it groups,
except for the new "strictFunctionTypes". This is a quick-fix, as the
tabris module should stay compatible to projects with "strict" flag.

Change-Id: I70f8eb963e4d224a43b741a5bf48077ac89bdf35