Add protected methods `_children`, `_find` and `_apply`
This allows overwriting the public `children` method to prevent access
to internals while the class itself can use the protected methods, which
are based on `_children`.

Add "protected" method handling in doc and typing generators. Adjust
input-tsx example to use this feature.

To simplify this change WidgetCollection now is iterable and can be
created without parameters.

Change-Id: I13300d5d921731dd50c7af4a864430f1fcc7fcd3