Add missing constructors in JSON definition files
Previously, the type of the `properties` parameter in the constructors
was given as `PropertiesObject`. The type definition generator replaced
that with the appropriate `<Widget>Properties` interface for each class.
For this reason, it was possible to omit constructor definition on some
Widgets and have the type definition generator generate the constructor
from the parent classes constructor. 
Since change 53028f8 "[TypeScript] Do not override `set` in every
NativeObject class", the `properties` parameter have type
`Properties<WidgetName>` where WidgetName is the class name of the
individual widget. The constructor cannot be inherited anymore and has
to be defined on every widget.

Change-Id: I21f88a9b4dd977477bd0bbd1b01a096a8cb673c4