Git-ignore npm's package-lock file
As of npm 5.0, each `npm install` creates or updates a file named
`packages-lock.json` that contains the complete list of installed
module versions. This file is meant to allow for reproducible builds.

For Tabris.js, there are only dev-dependencies, so the risk of
breakage is relatively low. Moreover, we always want to use the most
recent versions that match the ranges specified in the package.json.

Also, the nightly build should always use the latest versions
available in order to include the latest optimizations (e.g. to
JS minification) and to discover problems early.

Change-Id: Iadc7b951554b073da318cda129119a2b449b11ef