Add support for RefreshComposite
A RefreshComposite allows to use a pull-to-refresh gesture to provide
visual feedback while a long running operation is executing.

The widget mimics the refresh api of the `CollectionView` with the
properties `refreshEnabled`, `refreshIndicator` and `refreshMessage`.
It also provides the event `refresh`, which is fired when the user
triggers the refresh gesture.

Change-Id: Ic6ec8ad4bb584e5349de211d2fbf70e48a55b752