Remove "ts_ignore" check from type definition generator
The `ts_ignore` property for json definition files was introduced when a
"window.json" held definitions for the type "Timer", which was included
into the tabris namespace by the type definition generator.

Now, this type is defined in "timer.json" and the type definition
generator correctly generates globally scoped type definitions for it.

Until there is a good reason to have json definition files that should
be ignored by the type definition generator, support for `ts_ignore`
should be removed

Change-Id: I4175d653cfc4d969be17ed254d8dfd7d2d68a53c