Add support for "theme" property on NavigationBar
The theme defines the appearance used on the navigation bar. A `dark`
theme sets the foreground navigation icons to be of a light color,
whereas `light` sets the icons to a dark color. The theme should be set
in conjunction with the `background` property for contrast.

The value `default` selects the default theme that depends on the device
and on the app.

Available on Android 8+.

Change-Id: Ic215ea2f329912154ebe4dc99f436d8e054156a7