Add support for tabris.Printer
The tabris.Printer allows to print a PDF document using the native
printing capabilities of the device. The printer object is always
available via tabris.printer.

The printer offers the method print() which takes the data to print as
an ArrayBuffer or typed array. Additional parameter can be provided via
an options object.

The print method returns a promise which resolves to an event object
with the property `result`. The `result` can either be `completed` or
`canceled`. When printing fails the promise is rejected with an `Error`
parameter containing additional information about the error.

Change-Id: If9aad7c4995c7edf03b4fb920ab4bddb6fca85a4