Fire TabFolder selectionChanged on Tab dispose
Explicitly set the tab folder selection when the tab is removed from the
tab folder. This will ensure consistent behavior among platforms.
  * when closing a tab, which is not the most right one, set the
    selection to the right neighbor
  * when closing the most right tab, set it to the left neighbor
  * when closing the last tab, set it to null

Let selection getter return null when the TabFolder doesn't have any
children. This is the case when the last tab has been removed, since we
cannot explicitly SET null selection. The Tab is removed from the
TabFolder children in JS, but is not yet disposed, so the client would
return its cid instead of null in this case.

Fix #1579

Change-Id: Ia017b4b99ece188a4bdac6ed13f2b73311617040