Implement "appear" and "disappear" on Tab
Provide "appear" and "disappear" on Tab analogously to their Page
equivalents. This makes the Tab interface mostly compatible with Page
(with exception to the autoDispose flag on Page) and thus makes it
easier to exchange Page with Tab-based navigation and vice versa.

Change some of the tests to not use the calledWithMatch() method. When
there is a cyclic dependency between the TabFolder and the Tab
(introduced with storing the previous TabFolder selection),
calledWithMatch() on an object referencing both of them produces a stack
overflow.

Fix #815

Change-Id: I96293628cbe05aaccff0b73229c4639e695bc603