Introduce :root selector
The purpose of this selector is to allow the apply method to address the
widget that it is called on without needing a separate set method call.
It will become even more useful once we support child selectors.

:root behaves slightly different when used for the WidgetCollection
filter method where it always refers to tabris.ui. While this is very
unlikely to ever occur (you would have to explicitly construct a
WidgetCollection with tabris.ui in it), this is done for the sake of
consistency.

Documentation and examples to follow in a later commit.

Change-Id: I2061f5f5a459cd75ad20e668cc46d5b7fadca9e6