Add support for "selection" property on TextInput
The `selection` is a two element number array representing the text
selections start and end position. The native platform usually shows
selection handles so that the selection can be changed by the user.
A `selection` array where both numbers are the same represent a single
cursor at the given position.

The change also adds a corresponding event "select" which is fired when
the user edits the selection manually.

Change-Id: I63df13159dd1589933e5e1b2d6be97fdd8ddcb6f