Add support for Worker API
The Worker API allows to run JavaScript in parallel to the main
execution environment. The API follows the w3c spec for Worker support.

The implementation supports to create a new Worker, postMessages to it,
register for incoming onmessage and onerror events and to terminate the
worker.

See https://www.w3.org/TR/workers/
    https://developer.mozilla.org/en-US/docs/Web/API/Worker

Change-Id: I881a316076bfe39ab35c89e4755ebc468a201514