Remove "static" flag from CollectionView createCell and updateCell
These were marked as static, but in the implementation they can actually
be set.

This requires updates on the ts/jsx generators, since this now generates
change events for properties that (via function signatures) reference
"Widget". This did not work before since all references to widget in
declarations need to be prefixed with the "tabris" namespace. To
accomplish that "ts_type" can now be used to provide a type that does
this without also changing the generated documentation.

Change-Id: Ic639a611095be896da7c7f6cc70272bbcef087ed