Introduce negative tests for generated TypeScript declarations
This helps finding issues with "static" and "readonly" properties.

Parallel to "test.ts" files that have to compile without errors during
build, there are now also "fail.ts" files that have to output errors
during the build. The Errors expected are embedded in to the file itself
as a comment starting with "/*Expected". Every line in that comment is
treated as a required partial match.

Adjust some of the original test files to complement the new test
scenarios.

Change-Id: I9679e54074a253b75afd82650335ae1f78761d2d