Port Listeners API from tabris-decorators
This change simply ports the file "Listeners.ts" from
"tabris-decorators" release 2.5.1 to tabris core as "Listener.js", not
yet integrating it with the NativeObjects/widgets API of tabris.

There are minor changes beyond porting TypeScript to JavaScript:
* The methods "addListener", "removeListener", "once" and "trigger" now
return the event target, similarly to how they return "this" in
"Events.js".
* "once" is not re-implemented anymore
* Use Events.js to manage listeners in case the event target is not a
NativeObject instance
* Move event object initialization code to Events.js

To keep the behavior of Listeners.trigger, Events.trigger is slightly
revised. These changes are documented in the migration guide, except for
the promise error handling. This not a breaking change, just a warning
to prevent a common mistake that does not have to be documented
explicitly.

Change-Id: I1a7f25dcb9cc40116a5ae133c1bff10156c84133