[TypeScript] Allow converting any constructor to JSX component
By adding a factory to the prototype (for TypeScript support it can't be
the constructor itself, unfortunately), any class can be used as an JSX
element without having to extend the core framework. This will make it
easier for us and for users to introduce new JSX types (such as Popover)
or modify the behavior for specific subtypes.

With this change the existence of said factory is what marks a JSX
element, and no longer the jsxProperties property. That simplifies the
JSX generator a bit.

Adjust input-tsx example. Documentation to follow.

Change-Id: I559d33a653e78f54f8a6ba080ae30428e16ea588