Revise JSX children handling
JSX child elements (including text nodes) were previously handled
uniformly by convention without explicit runtime checks, type
declaration or documentation, or any way to adjust the behavior for
specific elements. These issus are resolved by this change as far as
possible.

At runtime children are now handled separately for each type by
overriding the jsx factory method. The JSX module was reworked to host a
new JsxProcessor class that provides helpers that the jsx factory
methods may access. A JSDoc annotation is used to establish
that these methods are executed in the context of an JsxProcessor
instance. By exchanging the instance in a specific scope the behavior of
these methods could now can be adjusted further.

The json schema for API documentation has been extended to provide
better type declarations and documentation:

The new `jsxChildren` property allows to define what type of children an
element accepts and is used to generate TypeScript declarations that
express this. The way TypeScript does this using the special
"ElementChildrenAttribute" Interface to assign one attribute
("children") to determine the type of supported child elements. At
runtime the property is supported as well, but that is currently not
documented. Unfortunately child type checks often do not work since all
our JSX elements are representing the "any" type. This is a due to a
TypeScript limitation tracked here:
https://github.com/Microsoft/TypeScript/issues/21699

Finally, the other new api-json property "jsxContentProperty" marks a
property as being settable via the JSX element content. This is used
only by the documentation generator. Currently this flag is always used
for string types, and the generator works under that assumption.

Change-Id: I02f313f4236fbf95353ef11324d8ca3d7f7e6a16