From now on the faq is a page like all the others in the documentation
Formerly the faq was processed in a special way to make it available independently
of a release version like "2.6.1" or "latest".
It was available under URL path /faq/index.html.
As we do not directly link to the FAQ anymore from anywhere, this change
moves the faq to "/latest/faq.html".
From now on it is also part of the table of contents (TOC).