Convert snippet "actionsheet.js" to JSX
Since this is the first use of JSX in our snippets there are some
further changes required:

In tasks.json enforce the file extension of the main module to be ".js".

Install eslint extension for react - it uses mostly the same logic as
Tabris.

Use the same eslint rules for snippets and framework code. Previously
this was not possible since snippets did not have a compilation step.

Fix resulting linter errors in snippets.

Change-Id: I95debaa2f24fda29b76633837759f14fcfa5a942