Introduce JSX element ActionSheetItem
ActionSheetItem should be used exclusively as an JSX element, as its
type is not entirely exchangeable with actions expressed as object
literals. This is due to the additional properties that enable JSX
support.

Enhance the json api schema to allow describing the alternative JSX type
of a given property.

Also fix return value of ActionSheet "actions" property. It should be a
safe copy with decoded image property.

Fix jsdoc in WidgetCollection.

Change-Id: I451e6ab237e9a11f35e1bcf9b50d6410c71e318c