Expose validation and normalization functions in fonts util
Change util-fonts to expose validation and normalization functions for
usage in an upcoming Font class. Improve naming of parseStyles function.

Fix a bug resulting in parsing invalid fonts without sizes not to throw
a human-readable error.

Enforce order of style and weight font parts according to CSS standard
and Tabris.js documentation. Prevously, order of style and weight was
arbitrary.

Change-Id: If0168400deff23784fb72926e9f0904af7b29c5d