Extend WidgetCollection API to allow access to host in filter functions
To make re-usable filter functions for use as selectors it may be useful
to have access to the widget that a widget collection was created from
(e.g. via children()).

This is done by giving the widget collection instance a filter is
applied to as the third parameter of the callback, just like with
standard JavaScript array filter functions. The collection itself has a
new property "host" that acts just like the ":host" selector, holding to
the widget in question.

Change-Id: I3bc6dcf4cea7824bc3641ec6f4b6ad912017c890