Introduce Constraint class
This class will be used later to represent left/right/top/bottom values
of layoutData. Currently it duplicates some logic of propertyTypes.js,
this will be resolved when the class is integrated with layoutData.

No type guards for this type since it is not useful for custom component
development.

Type declarations/documentation to follow separately.

Change-Id: I2b72b060d5ffe99caf11c68e7b23078651ad0ea4