Turn former Layout object in to a class
The new class handles all layout specifics, composites only need to
register and de-register themselves via the public add/remove methods.

LayoutQueue is extracted as a separate singleton so multiple instances
of Layout does not pose a problem later. Currently composite-extending
widgets that support layoutData get the same "default" instance of this
new class set as the "layout" property. No documentation yet.

There is a slight change in behavior towards the client in that
layoutData is rendered even if it was never set, as it now has a default
"all-auto" value. This is irrelevant in real applications since nearly
all widgets have layoutData, but causes some changes in numerous tests.

Also, avoid "calledWithMatch" expects with widgets since this
can cause stack overflows with the layout property.

Change-Id: Ib8ad544db45fcde2bdd43d298a4b9fa05240da06