Remove Ui widget
Objects formerly found on the ui object are now exported directly on the
tabris object, just like other singletons.

Turn the Tabris class in to a NativeObject, so the association between
itself and the drawer/contentView instances mirrors those of Popover and
its own contentView. Make Drawer extend ContentView for consistency.

Change-Id: I1241c49a896c1d20a2ab0249788506ea7bac88f6