Implement custom Widget toString method
Returns a pseudo selector string (currently not supported as an actual
selector in tabris) that includes type, cid, id, and classList.

Also override the toStringTag property used by the default toString
implementation. This prevents widgets to be stringified as "[Object
object]" in some rare cases. It will be "[object <widgetType>]" instead.

Change-Id: I30b66198a13ce6d4fd7fc37c2e8b4e556ed8c704