Remove ColorShader type
Currently, the Shade types are responsible for encapsulating protocol
formats and to validate input.

Since the input validations are to be done by public types and protocol
format is already encapsulated in property-types.js, Shader types don't
have any responsibility now. Remove the ColorShader type.
LinearGradientShader will be removed when a LinearGradient class is
introduced in the future.

Change-Id: If25ce15b9c6b7c682897ba16f0b4b82461a2d03d