Add JSX support to AlertDialog
Follow the same patterns established with ActionSheet.

Add "texts" array to close event, since the TextInput elements are
disposed after the dialog closes and can't be read anymore. That was not
very async-api friendly.

Also allow AlertDialog.open to be called with just a string. This is an
approximation of the browser "alert" method and may proof very useful as
an alternative to console.log in snippets.

Change-Id: Iafe8bb314356c794be0983e9f6f7ad70d08d078f