Fix console snippet trace example for iOS
An oddity of iOS stack traces is that they don't include arrow functions
without curly braces, resulting in an effectively empty stack trace in
this specific example.

Very likely there is nothing we can do about this except printing the
original trace, which is what happened in this snippet. This is unlikely
to happen in a real-world app.

Modify the snippet to use curly braces to create a correct stack trace.

Change-Id: Ia3402f943d313cbc1598b440f379115911fede84