Refactor and unify native object creation
Rename NativeObject#_create to _nativeCreate to match _nativeSet, etc.
Call this method in the NativeObject constructor, not in arbitrary
subclasses. Let it also handle registration and initial properties so
that this behavior can be easily adjusted in subclasses. Overriding
_nativeType is enforced.

Fix/add some JsDoc where it touches on native object creation.

Change-Id: I1d7c47b112e062ab089f7ad6a3a5a2d60f3417ab