Simplify set method and constructor type declarations
Instead of overriding the set method on each new subclass of set, modify
the generic type "Properties<T>" so that it can derive the properties
parameter from `this` and use it on NativeObject#set. Can also be used
for constructor parameters.

Methods and readonly properties are stripped. This is possible approach
because of the previous commit and inspired by
https://stackoverflow.com/a/52473108.

The upside is approach that a custom component author gets better
out-of-the-box auto-completion for `set` and a much simpler constructor
signature. It also simplifies our generator code.

The down side is that in rare cases (CollectionView, Picker) properties
have a function type and should not be stripped. The set method needs to
be overwritten then.

Change-Id: I4c12f9e3ff6328fe4b454c427bcb1b402f47e0bd